(bug:3260712) Fixed the NET and XMLBIF IO when saving SSBN networks. I also tested regular BNs to make sure it did not stop working for these cases that were already working before this fix. The problem was with the order of the nodes, which were a bit different when constructing SSBN. I had to fix them to use the CPT node order.
I only fixed the XMLBIF version 0.6. I did not fix previous versions.

I changed the parent index to start at 0, as it was before in previous versions. This way it is still compatible with previous versions. I.e., it does not start at 1, as I changed earlier when fixing this bug.