Fixed a very minor bug in the test suite, which was always verifying changes in the expected assets assuming that they really changed, but such assumption is not right, because probabilities may not always change after propagation.
Also, fixed a bug on the method for exiting from a question. The bug was causing the engine to think that the assets were zero when exiting from a disconnected question, because it was trying to calculate something like marginal asset, and such thing as marginal asset is not defined, hence it was returning zero assets.