Made changes in MEBNController because the last commit was forcing the SSBNCompilationPane to be always displayed and the evidence tree to be updated, regardless of how the ssbn generator was implemented.
Before the previous commit, SSBN generators were instances of IMediatorAwareSSBNGenerator, and each generator was overwriting the method showSSBN in order to show the SSBN in a customized manner.
 
After the previous commit, showSSBN was moved to MEBNController, so the SSBN generators were not able to customize the way they show the SSBN (well, it's possible to customize, but SSBNCompilationPane was always shown, and the evidence tree was always updated -- which caused errors when evidence tree is not populated).

Now, MEBNController will only show SSBNCompilationPane if a boolean attribute is active (previously, this boolean attribute was overwritten by the MEBNController, so the attribute was quite useless) and the evidence tree will be updated only if the SSBNCompilationPane is actually shown.

Note: other implementations of SSBN generators MUST be revised because of the fact that showSSBN was migrated to MEBNController and LaskeySSBNAlgorithm is not invoking it anymore (so classes extending LaskeySSBNAlgorithm and overwriting showSSBN are affected). As far as I know, plug-ins that probably needs revisions are: MEDG, Multiple-Sectioned SSBN, and Continuous MEBN.