Fixed bug in TxtSaver that was causing cells of data tables of discretized attributes to contain numbers (index of states) instead of name. This was virtually causing all discretized attributes to remain numeric (because values in saved files will be index numbers).
Additionally, unbminer preprocessor will ask to save discretized data as compressed file or not.

Also fixed minor bug in range discretization with zero, which was not considering the max value in the range if data did not contain zeros.

Discretization methods with zeros will now append prefixes before state names, so that learning modules don't consider states starting with numbers as numbers.